add message

+++ main/views.py
@@ -1,32 +1,101 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
 
 
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user=user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password incorrect')
+
+    context = {'page': page}
+    return render(request, 'main/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerUser(request):
+    page = 'register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    context = {'page': page, 'form': form}
+    return render(request, 'main/login_register.html', context)
+
+
 def home(request):
-    q = request.GET.get("q") if request.GET.get("q") != None else ''
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
         Q(description__icontains=q))
     topics = Topic.objects.all()
+    room_count = rooms.count()
 
-    context = {'rooms': rooms, 'topics': topics}
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'main/home.html', context)
 
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all()
 
-    context = {'room': room}
+    if request.method == 'POST':
+        body = request.POST.get('body')
+        message = Message.objects.create(
+            user=request.user, room=room, body=body)
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'room_messages': room_messages}
     return render(request, 'main/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+
+    if request.user != room.host:
+        return HttpResponse('No permission to create')
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -37,9 +106,14 @@ def createRoom(request):
     return render(request, 'main/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('No permission to update')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -50,8 +124,12 @@ def updateRoom(request, pk):
     return render(request, 'main/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('No permission to delete')
 
     if request.method == 'POST':
         room.delete()
